requirements.txt
matplotlib
universal_usbtmc
pyserial
